move functionality up to base to prep for mongo key store
